fix(tts-service): fix fp16 overflow on low_vram tier causing synthesis failure
LOW_VRAM tier used fp16 which overflows to inf/nan on older CUDA GPUs
(pre-Volta, no native Tensor Core fp16 support), causing every synthesis
call to fail with "probability tensor contains either inf, nan or element
< 0".

- Change LOW_VRAM tts_precision from fp16 to fp32; the 0.6B model in
  fp32 uses ~1.2 GB VRAM, well within any <10 GB GPU budget
- Add runtime fp16→fp32 fallback in TTSEngine.synthesize_to_file: if
  generate_voice_clone raises the NaN/inf RuntimeError on any tier, cast
  the model to fp32, invalidate the cached voice prompt (computed under
  fp16 weights), recompute, and retry — guards against edge cases on
  hardware that reports fp16 support but produces unstable results

+++ tts-service/app/core/hardware.py
@@ -101,7 +101,7 @@ _TIER_CONFIGS: dict[HardwareTier, EngineConfig] = {
         tier=HardwareTier.LOW_VRAM,
         tts_model='Qwen/Qwen3-TTS-12Hz-0.6B-Base',
         tts_device='cuda',
-        tts_precision='fp16',
+        tts_precision='fp32',  # fp16 overflows on older GPUs → inf/nan logits; 0.6B fp32 = ~1.2 GB VRAM
         llm_model='qwen3:4b',
         narration_strategy='chunked',
         narration_chunk_words=1000,
